+++ warp/__init__.py
@@ -333,6 +333,13 @@ from warp._src.context import is_peer_access_supported as is_peer_access_support
 from warp._src.context import is_peer_access_enabled as is_peer_access_enabled
 from warp._src.context import set_peer_access_enabled as set_peer_access_enabled
 
+from warp._src.context import Allocator as Allocator
+from warp._src.context import get_device_allocator as get_device_allocator
+from warp._src.context import set_cuda_allocator as set_cuda_allocator
+from warp._src.context import set_device_allocator as set_device_allocator
+from warp._src.utils import ScopedAllocator as ScopedAllocator
+from warp._src.rmm_allocator import RmmAllocator as RmmAllocator
+
 
 # category: CUDA Graph Management